simplify abstractions
