+++ src/internal/origin_directory_converter.js
